# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /admin/payouts:
    get:
      operationId: AdminPayoutController_list
      parameters:
      - name: status
        required: false
        in: query
        schema:
          type: string
      - name: organizerId
        required: false
        in: query
        schema:
          type: string
      - name: from
        required: false
        in: query
        schema:
          type: string
      - name: to
        required: false
        in: query
        schema:
          type: string
      - name: page
        required: false
        in: query
        schema:
          type: string
      - name: limit
        required: false
        in: query
        schema:
          type: string
      responses:
        '200':
          description: ''
      summary: List payouts with filters
      tags:
      - Admin / Payouts
...
  /admin/payouts/{id}/approve:
    post:
      operationId: AdminPayoutController_approve
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payout'
      summary: Approve a payout for processing
      tags:
      - Admin / Payouts
  /admin/payouts/{id}/mark-settled:
    post:
      operationId: AdminPayoutController_markSettled
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payout'
      summary: Mark a payout as settled (manual transfer)
      tags:
      - Admin / Payouts
...
  /admin/pricing/campaigns:
    get:
      operationId: AdminPricingCampaignController_list
      parameters: []
      responses:
        '200':
          description: ''
      summary: List pricing campaigns with status
      tags:
      - Admin / Pricing
    post:
      operationId: AdminPricingCampaignController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCampaignDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricingCampaign'
      summary: Create a pricing campaign
      tags:
      - Admin / Pricing
  /admin/pricing/campaigns/{id}:
    patch:
      operationId: AdminPricingCampaignController_update
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCampaignDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricingCampaign'
      summary: Update or cancel a pricing campaign
      tags:
      - Admin / Pricing
...
  /admin/referral/settings:
    get:
      operationId: AdminReferralTierController_getSettings
      parameters: []
      responses:
        '200':
          description: ''
      summary: Get referral global settings
      tags:
      - Admin / Referrals
    patch:
      operationId: AdminReferralTierController_updateSettings
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReferralSettingsDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
      summary: Update referral global settings
      tags:
      - Admin / Referrals
  /admin/referral/tiers:
    get:
      operationId: AdminReferralTierController_listTiers
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReferralTierConfig'
      summary: List all referral tiers
      tags:
      - Admin / Referrals
    post:
      operationId: AdminReferralTierController_createTier
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTierDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReferralTierConfig'
      summary: Create a referral tier
      tags:
      - Admin / Referrals
  /admin/referral/tiers/{id}:
    patch:
      operationId: AdminReferralTierController_updateTier
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTierDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReferralTierConfig'
      summary: Update a referral tier
      tags:
      - Admin / Referrals
    delete:
      operationId: AdminReferralTierController_deleteTier
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '200':
          description: ''
      summary: Delete a referral tier
      tags:
      - Admin / Referrals
...
  /admin/weather/alerts:
    get:
      operationId: AdminWeatherAlertController_list
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WeatherAlert'
      summary: List all weather alerts
      tags:
      - Admin / Weather
    post:
      operationId: AdminWeatherAlertController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlertDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeatherAlert'
      summary: Create a weather alert
      tags:
      - Admin / Weather
  /admin/weather/alerts/{id}:
    delete:
      operationId: AdminWeatherAlertController_expire
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '200':
          description: ''
      summary: Expire a weather alert early
      tags:
      - Admin / Weather
...
info:
  title: Offbeat प्रवासी API
  description: Full-scale, production-ready trek discovery and booking platform.
  version: 1.0.0
  contact: {}
...
servers:
- url: http://localhost:4000
  description: Local development
- url: https://api.offbeatpravasi.com
  description: Production
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
    ApiKey:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
...
    CreateAlertDto:
      type: object
      properties: {}
...
    CreateCampaignDto:
      type: object
      properties: {}
...
    CreateTierDto:
      type: object
      properties: {}
...
    Payout:
      type: object
      properties:
        id:
          type: string
        organizerId:
          type: string
        amount:
          type: number
        platformFee:
          type: number
        netAmount:
          type: number
        status:
          enum:
          - PENDING
          - APPROVED
          - PROCESSED
          - SETTLED
          - FAILED
          type: string
        periodStart:
          format: date-time
          type: string
        periodEnd:
          format: date-time
          type: string
        settledAt:
          format: date-time
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
      - id
      - organizerId
      - amount
      - platformFee
      - netAmount
      - status
      - periodStart
      - periodEnd
      - createdAt
      - updatedAt
...
    PricingCampaign:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        trekIds:
          type: array
          items:
            type: string
        discountType:
          enum:
          - PERCENTAGE
          - FLAT
          type: string
        discountValue:
          type: number
        maxCap:
          type: number
          nullable: true
        minBookingAmount:
          type: number
          nullable: true
        startDate:
          format: date-time
          type: string
        endDate:
          format: date-time
          type: string
        isActive:
          type: boolean
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
      - id
      - name
      - trekIds
      - discountType
      - discountValue
      - startDate
      - endDate
      - isActive
      - createdAt
      - updatedAt
...
    ReferralTierConfig:
      type: object
      properties:
        id:
          type: string
        tier:
          enum:
          - BASE
          - SILVER
          - GOLD
          type: string
        minSuccessfulReferrals:
          type: number
        rewardPerReferralInr:
          type: number
        refereeDiscountInr:
          type: number
      required:
      - id
      - tier
      - minSuccessfulReferrals
      - rewardPerReferralInr
      - refereeDiscountInr
...
    UpdateCampaignDto:
      type: object
      properties: {}
...
    UpdateReferralSettingsDto:
      type: object
      properties: {}
...
    UpdateTierDto:
      type: object
      properties: {}
...
    WeatherAlert:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        body:
          type: string
        severity:
          enum:
          - ADVISORY
          - WATCH
          - WARNING
          type: string
        affectedRegion:
          type: object
          nullable: true
        isActive:
          type: boolean
        expiresAt:
          format: date-time
          type: string
          nullable: true
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
      - id
      - title
      - body
      - severity
      - isActive
      - createdAt
      - updatedAt
...
security:
- ApiKey: []